Added a new feature! AddBlock detection :+1:
A new java file is created, with a aux var inside. This java file is named ads.
If addblock is activated, ads.js will not load and the var will be undefinded!

for more info about this, look up this link:
http://stackoverflow.com/questions/4869154/how-to-detect-adblock-on-my-website

Apart from that, I moved the input card a little bit lower, so the title GLORIO can be read properly.